chore(wasmbinding): delete CustomQuerier since we have QueryRequest::Stargate now
