+++ wasmbinding/bindings/marshalling_test.go
@@ -12,7 +12,6 @@ import (
 
 	"github.com/NibiruChain/nibiru/app"
 	"github.com/NibiruChain/nibiru/wasmbinding/bindings"
-	"github.com/NibiruChain/nibiru/x/common/testutil/genesis"
 )
 
 type TestSuiteBindingJsonTypes struct {
@@ -36,57 +35,6 @@ func (s *TestSuiteBindingJsonTypes) SetupSuite() {
 	s.fileJson = fileJson
 }
 
-func (s *TestSuiteBindingJsonTypes) TestQueries() {
-	testCaseMap := map[string]any{
-		"all_markets":      new(bindings.AllMarketsResponse),
-		"reserves":         new(bindings.ReservesResponse),
-		"base_price":       new(bindings.BasePriceResponse),
-		"position":         new(bindings.PositionResponse),
-		"positions":        new(bindings.PositionsResponse),
-		"module_params":    new(bindings.PerpParamsResponse),
-		"premium_fraction": new(bindings.PremiumFractionResponse),
-		"metrics":          new(bindings.MetricsResponse),
-		"module_accounts":  new(bindings.ModuleAccountsResponse),
-		"oracle_prices":    new(bindings.OraclePricesResponse),
-	}
-
-	for name, cwRespPtr := range testCaseMap {
-		s.T().Run(name, func(t *testing.T) {
-			err := json.Unmarshal(s.fileJson[name], cwRespPtr)
-			s.Assert().NoErrorf(err, "name: %v", name)
-			jsonBz, err := json.Marshal(cwRespPtr)
-			s.NoErrorf(err, "jsonBz: %s", jsonBz)
-		})
-	}
-}
-
-func (s *TestSuiteBindingJsonTypes) TestToAppMarket() {
-	var lastCwMarket bindings.Market
-	for _, ammMarket := range genesis.START_MARKETS {
-		dummyBlockHeight := int64(1)
-		cwMarket := bindings.NewMarket(
-			ammMarket.Market,
-			ammMarket.Amm,
-			"index price",
-			ammMarket.Amm.InstMarkPrice().String(),
-			dummyBlockHeight,
-		)
-
-		// Test the ToAppMarket fn
-		gotAppMarket, err := cwMarket.ToAppMarket()
-		s.Assert().NoError(err)
-		s.Assert().EqualValues(ammMarket.Market, gotAppMarket)
-
-		lastCwMarket = cwMarket
-	}
-
-	// Test failure case
-	sadCwMarket := lastCwMarket
-	sadCwMarket.Pair = "ftt:ust:xxx-yyy!!!"
-	_, err := sadCwMarket.ToAppMarket()
-	s.Error(err)
-}
-
 func getFileJson(t *testing.T) (fileJson map[string]json.RawMessage) {
 	file, err := os.Open("execute_msg.json")
 	require.NoError(t, err)
@@ -102,13 +50,7 @@ func (s *TestSuiteBindingJsonTypes) TestExecuteMsgs() {
 	fileJson := getFileJson(t)
 
 	testCaseMap := []string{
-		"market_order",
-		"close_position",
-		"add_margin",
-		"remove_margin",
 		"donate_to_insurance_fund",
-		"peg_shift",
-		"depth_shift",
 		"edit_oracle_params",
 		"set_market_enabled",
 		"insurance_fund_withdraw",
