+++ wasmbinding/exec_perp.go
@@ -20,123 +20,6 @@ func (exec *ExecutorPerp) MsgServer() perpv2types.MsgServer {
 	return perpv2keeper.NewMsgServerImpl(exec.PerpV2)
 }
 
-func (exec *ExecutorPerp) MarketOrder(
-	cwMsg *bindings.MarketOrder, sender sdk.AccAddress, ctx sdk.Context,
-) (
-	sdkResp *perpv2types.MsgMarketOrderResponse, err error,
-) {
-	if cwMsg == nil {
-		return sdkResp, wasmvmtypes.InvalidRequest{Err: "null open position msg"}
-	}
-
-	pair, err := asset.TryNewPair(cwMsg.Pair)
-	if err != nil {
-		return sdkResp, err
-	}
-
-	var side perpv2types.Direction
-	if cwMsg.IsLong {
-		side = perpv2types.Direction_LONG
-	} else {
-		side = perpv2types.Direction_SHORT
-	}
-
-	sdkMsg := &perpv2types.MsgMarketOrder{
-		Sender:               sender.String(),
-		Pair:                 pair,
-		Side:                 side,
-		QuoteAssetAmount:     cwMsg.QuoteAmount,
-		Leverage:             cwMsg.Leverage,
-		BaseAssetAmountLimit: cwMsg.BaseAmountLimit,
-	}
-	if err := sdkMsg.ValidateBasic(); err != nil {
-		return sdkResp, err
-	}
-
-	goCtx := sdk.WrapSDKContext(ctx)
-	return exec.MsgServer().MarketOrder(goCtx, sdkMsg)
-}
-
-func (exec *ExecutorPerp) ClosePosition(
-	cwMsg *bindings.ClosePosition, sender sdk.AccAddress, ctx sdk.Context,
-) (
-	sdkResp *perpv2types.MsgClosePositionResponse, err error,
-) {
-	if cwMsg == nil {
-		return sdkResp, wasmvmtypes.InvalidRequest{Err: "null close position msg"}
-	}
-
-	pair, err := asset.TryNewPair(cwMsg.Pair)
-	if err != nil {
-		return sdkResp, err
-	}
-
-	sdkMsg := &perpv2types.MsgClosePosition{
-		Sender: sender.String(),
-		Pair:   pair,
-	}
-	if err := sdkMsg.ValidateBasic(); err != nil {
-		return sdkResp, err
-	}
-
-	goCtx := sdk.WrapSDKContext(ctx)
-	return exec.MsgServer().ClosePosition(goCtx, sdkMsg)
-}
-
-func (exec *ExecutorPerp) AddMargin(
-	cwMsg *bindings.AddMargin, sender sdk.AccAddress, ctx sdk.Context,
-) (
-	sdkResp *perpv2types.MsgAddMarginResponse, err error,
-) {
-	if cwMsg == nil {
-		return sdkResp, wasmvmtypes.InvalidRequest{Err: "null add margin msg"}
-	}
-
-	pair, err := asset.TryNewPair(cwMsg.Pair)
-	if err != nil {
-		return sdkResp, err
-	}
-
-	sdkMsg := &perpv2types.MsgAddMargin{
-		Sender: sender.String(),
-		Pair:   pair,
-		Margin: cwMsg.Margin,
-	}
-	if err := sdkMsg.ValidateBasic(); err != nil {
-		return sdkResp, err
-	}
-
-	goCtx := sdk.WrapSDKContext(ctx)
-	return exec.MsgServer().AddMargin(goCtx, sdkMsg)
-}
-
-func (exec *ExecutorPerp) RemoveMargin(
-	cwMsg *bindings.RemoveMargin, sender sdk.AccAddress, ctx sdk.Context,
-) (
-	sdkResp *perpv2types.MsgRemoveMarginResponse, err error,
-) {
-	if cwMsg == nil {
-		return sdkResp, wasmvmtypes.InvalidRequest{Err: "null remove margin msg"}
-	}
-
-	pair, err := asset.TryNewPair(cwMsg.Pair)
-	if err != nil {
-		return sdkResp, err
-	}
-
-	sdkMsg := &perpv2types.MsgRemoveMargin{
-		Sender: sender.String(),
-		Pair:   pair,
-		Margin: cwMsg.Margin,
-	}
-	if err := sdkMsg.ValidateBasic(); err != nil {
-		return sdkResp, err
-	}
-
-	goCtx := sdk.WrapSDKContext(ctx)
-	return exec.MsgServer().RemoveMargin(goCtx, sdkMsg)
-}
-
 func (exec *ExecutorPerp) InsuranceFundWithdraw(
 	cwMsg *bindings.InsuranceFundWithdraw, ctx sdk.Context,
 ) (err error) {
